LayoutTests/imported/w3c/web-platform-tests/wasm/jsapi/table/set-reftypes.tentative.any-expected.txt


FAIL Arguments for anyfunc table set assert_equals: expected (function) function "function 0() {
    [native code]
}" but got (object) null
FAIL Arguments for externref table set assert_equals: expected object "[object Object]" but got null
